+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Erase discarded reasoning block in-place
 - Restrict a hidden-line notice to one row
 - Use fast instead of ultrafast with Codex
+- Heed X-Should-Retry header's wishes
 
 ## [0.8.0] - 2026-09-20